Added barebones UI for index.tsx (Home) page. Currently non-functional.
List of existing activities is represented by a hardcoded array. This will hopefully be replaced with items from the
database once that's set up. It's implemented using a FlatList component, which should allow scrolling but is not
at the moment. I will attempt to debug at a later date. All components are styled with a stylesheet at the bottom
of the page, which should hopefully make refactoring easier later on. Buttons are also made using a custom component
for the same reason.