Added activity page
Added activityPage, which will be able to display information based off of the habit name passed to it by the navigation
button in index.tsx. Currently it displays associated images in a FlatList component. Pending data visualization and styling.
Images display right now using a link that is hosted on Discord (This should be changed later to use images stored locally
or on FireBase).