Line graph added to activityPage.tsx
Graph renders in Android Studio Emulator, not in web as of last testing. Made using an amalgam of d3.js and svg.
Current mock data used does not reflect days of the week or actual dates. Should be able to modify that fairly easily.
Axes are unlabeled for not, but are numbered.